Train, Player, and Camera Movement
Created a simple train engine model to see what it will look like in game
for sizing.

Set some waypoints in the world.
Made a path script that organizes the waypoints. The train uses the path
class to find the nexxt waypoint to move to when it reaches the current
one.
Train moves straight towards the points at a constant speed without
rotating.

The camera follows the train's y position, but not its x. This is so the
train can move from the left to the right side of the screen.

The player can move around on the train while it's moving. This works by
having the player be a child of the train. I'm not 100% sure how this will
interact with the train and player both having rigidbodies. I feel like
I've had trouble with these interactions before. Currently it works and
lets the player move where the train is moving, but also lets the player
change their position on the train.