Cars Move Independently
All the train cars now follow the tracks.
Each car has the Train script.
The movespeed of the trains is static so they all have the same speed.
Also when the engine reaches the end of the track, the whole train stops.
The subsequent cars don't drive into the engine.

When the player hops trains, it also changes what car is its parent (to
the train its on) so that the player follows the correct train's movement,
not the engine.

Made the track longer. The corners don't seem to need to be especially
smooth. The way the wheels work, it tends to smooth out the corners a bit
so that it's not super noticeable.