Enemies, Health, Rail Changes
Path
Started to create some methods to change the path a bit.
One method to add points to the path and space out the points so they are
equal distance apart. Nothing done yet.
Another method started to create a mesh to put on the path to draw the
rails as a texture

Player
Uses the new health script to track its hp. Should die and respawn after a
few seconds when hp goes to 0, but nothing does damage yet and the
spawnpoint isn't set either.

Train
Helper method that can approximate where the train wil be in x seconds.
Uses the current speed of the train so it won't be accurate when the speed
changes.
It also doesn't return the exact spot the train will be, instead it
returns the position of the point the train will be moving towards at that
point.

Enemy
Created an enemy script. It can move towards a target. Attempts to
intercept the train when it spawns.