Respawning Enemies and Float Health
Enemies can be set to respawn. After they die, 3 seconds later they will
respawn where they first spawned.
If there is no train engine set for the enemies, they just stand still.
This is going to be temporary for using them as target dummies.

health is now a float instead of an int. May go back and change it back to
an int at some point. I just made the damage for the guns be a float so
health is now a float too.

Made the enemy into a prefab and coloure it red.
Made the bullet projectile larger.