Hit Particles and Shoot Sounds
Grabbed some shooting sounds from a different project (Bullet Farm)
Added a shoot sound to the gun. It's ok for now, but will probably need an
upgrade.

Created a blood splatter particle effect.
It's spawned by the bullet when it hits an enemy.
Tried having the splatter travel in the same direction the bullet was
travelling, but having the splatter move in the direction straight away
from the center of the enemy hit towards where the bullet impacted looks
much better.
The particles are just red squares that shrink and slow down over time.

ParticleController
It just deletes the particles after half a second.