Aim Preview
Added a line from the gun to the player's mouse to show the player where
they're aiming.
The Player class now tells the current weapon where the mouse is aiming

RangedWeapon
Added in side-to-side recoil to the weapon. This makes the weapon have
more spread once its been firing for a while. The weapon changes its
rotation on the y axis so bullets fire in a different direction. This
rotation is randomized within the bounds given by the current recoil
percent. Higher recoil has a larger bounds.

Added a sound effect to getting hit.

Bullets travel twice as fast