Melee Weapons, Enemy Combat, Controller Aiming with Lock On
Enemy
Enemy will chase targets when they get in its aggro range. Once it gets
close (inside attackRange), it will attempt to make a melee attack.
The enemy uses the weapon class to attack. It has a melee weapon
When the enemy dies, it stops moving
It also has an OnDeath event for others to subscribe to to be notified
when it dies.

Player
Added support for controller aiming. Player can move the cursor around
with a controller
When the player dies, they can't move. Gravity is turned off so they don't
fall through the map while trying to respawn.

MeleeWeapon
Contains the funtionality for a melee weapon.
It just turns on and off a given collider

EnemyAI
Some mocked up methods that might be useful for AI
Does nothing at the moment. AI functionality is in the Enemy class

ThumbCursor
Controls the movement of the cursor with the right thumbstick on a
controller.
If a player claims the cursor, the cursor will move around relative to the
player.
The cursor will snap to enemies. Once the attached trigger sphere hits an
enemy, the cursor locks on to it.
The player can use the cursor's position for aiming.
While locked on to an enemy, the cursor will follow the enemy as it moves.
Moving the cursor off of the enemy will unattach it from the enemy.
While locked on to an enemy, the player will aim straight towards the
enemy, not the cursor anymore.

Problems
Only tested in the flat Shooting Gallery world. May or may not work when
the player is above the enemies like on the main train scene.
If the ControllerAiming bool is true, the player tries to use the
controller aiming. It tries to check to make sure a controller is plugged
in. If one isn't, it should default to mouse aiming.
However, this seems a little inconsistent. Once a controller is plugged
in, checking if a controller is there will be true even after the
controller has been unplugged. i.e. Once a controller has been plugged in
once, it always thinks there is one. This may only be a problem in the
editor.