Player Aiming, Clean Up Train Scene
With the locking on that aiming with a controller adds, the player would
try to face straight towards the target.
Fixed so the player only rotates around the up vector.

Made the track a bit longer for better testing
Increased the size of the ground plane
Removed some unnnessecary cubes attached to the train and the wheels of
the train