Ranged Weapon Aiming
Ranged weapons now try to orient themselves towards the aimpoint given to
them by the player. The gun is still affected by recoil changing its
orientation, it just tries to move to point towards the aim point when it
has no recoil.

Added an empty parent object to store all the train objects in. Cleans up
the scene a bit. The train appears to still work. Each car follows the
track like it's supposed to.