Automatic Rail Curves
Created a system to calculate the curve of the rail given a max curve
angle.
This is instead of having to create the curves manually. Now whenever I
place waypoints, they can just be straight from point A to B and the track
will curve to fit.

The algorithm isn't implemented into the track creation yet. It just works
for some testing points at the moment.
It does seem to work for all orientations of the rail (straight to left,
straight to left, angled to straight).