Track Auto Curve Creation
The track now automatically creates the curves instead of having to make
them manualy.

First, given transforms get turned into a list of points. A list because
each curve will have a different number of new points being created that
need to be inserted where the original corner point was.
For each point, calculate the curve around it.
The points that make up the curve get inserted into the list and the
original corner is removed.
The list is then converted to an array and passed into the even spacin
function.
This function was modified to take in an array of vector3's instead of
using the array of transforms. It still operates on the evenPoints array
directly instead of returning its own array.