Ranged Enemy, Faster Movement, Healthbar
Increased the movespeed for the player and the aiming cursor.
Moved the camera out a bit.

Added a healthbar image. The Health class controls where it is and how big
the healthbar is.
The healthbar just keeps its original rotation, it doesn't look at the
camera. It is also unattached to the gameobject with the health component
so many healthbars would take up a lot of space in the hierarchy. When you
lose health, the healthbar loses ticks to show losing health, but the
discrete ticks don't match the amount of damage enemies do (hp bar has 5
ticks so each is 20 hp. Enemies do 10 damage so you lose half a tick which
looks weird). The image stays centered as it gets smaller which isn't what
I want. It should stay left-aligned as the health bar empties.
Health script has numbers for the current image hard coded in. This is
just temporary to test the tiled draw mode style of healthbars
(SpriteRenderer->Draw Mode->Tiled). Not super
in love with it so it may change.

Added a ranged enemy.
It's just a regular enemy, but with its melee weapon swapped out for a
ranged weapon. And changed aggro and attack ranges. Had to make the enemy tell its weapon where it's aiming to
make them aim the gun, but it seems to work otherwise.
Updated the projectile class to take a hitTag. This tells the bullet what
tag it can hit. The player's bullets hit "Enemy" and the enemy bullets hit
"Player".

Removed some unneeded comments and variables from the train.