Reworked Train Interception
Changed the logic of the train checking where it will be in x seconds.
Before, the train added up all the distances between the points ahead of
it.
Now, it leverages the Path class' knowledge of the distance between
points.
The path calculates how many points fit in a given distance and then
returns a point that is that many points ahead of the given location.

Still returns a location that is ahead of where the train will be at that
time. It should somehow give the enemy a range of different positions
between the front of the train and the back.