Train Interception
Made a better train interception method.
Not made yet, just planning.
The intention is to allow it to return multiple different points where and
agent could path to board the train instead of just returning one point in
front of the train.
The technique in the plan is to make x boarding points on each car where
enemies can enter from the ground. These boarding points would be a pair
of points. One on the the train and one on the ground. If the enemy makes
it to the point on the ground, they can use to it to travel to the point
on te train (by climing or jumping from there).