Boarding Links and Boarding Lookahead
BoardingLink
Class to store the link between the ground position and position on the
train. Enemies will be able to go to the ground position to be able to
board the train.
Currently just calculates the position the ground point will be at given a
certain train location and rotation

Train
Made GetBoardingLocationsInTime(t) functional.
Calculates where the train will be in time t and at what rotation.
It then passes this information to the attached BoardingLinks to find
their ground positions.
Then it returns these ground positions for the enemy to use them to plan a
path.

Problems:
The lookahead doesn't work properly until the front and back wheels of the
train are on the track.
Each train calls its own links. An enemy would therefore need to check
each train for the links. This may or may not be a bad thing.