Enemies Board Trains
Enemies can now look for where the train's boarding points will be and
calculate if they can get to them in time.
If they can, they move towards that point.
When they collide with a ground point of the boarding link (collide with
any, not necessarily the one they were trying to hit), they teleport to
the linked on board location.
When on the train, the enemy stops moving so they don't immediately run
off the train.
When they collide with the train, they make the train their parent so they
move with it (like the player does).

BoardingLink
Now has a collider on the ground point.

Player
use CompareTag() insead of == because it's faster/more efficient and
consistent with other places in the code.

Train
Used SignedAngle to find the angle instead of Angle. The boarding links
need the sign of the angle to be orientated properly.