Boarding Callback
The train now checks if the rear wheels are on the track before it lets
enemies board (because the calculations are wrong otherwise).
Enemies try to board, the train says no, but the enemy sent a method to
call when the train is ready. The train adds this method to an event that
is called when the back wheels enter the track. So the enemies are
notified right when the train is able to be boarded.

Added some better debugging output. Things state their name so it's a bit
easier to tell what is happening to each object.