Enemies Attacking on Train
Enemy
When enemies first teleport on the train, they try to find targets (only
the player for now). Then they try to move towards the player.
They check if they are in range of their target to attack by raycasting.
Used to be distance center to center, but that doesn't work with the train
and its rectangular hitboxes.
Enemies can also attack the train and do damge to it (but nothing happens
when the train dies and they won't because they find the player).
The raycasting is slightly wrong now. It checks if it hit something on the
Train layer or the Damageable layer (player) to do damage. It shouldn't
hit what it isn't currently searching for.

Boarding the train is still being tested so it has some debugging
behaviour (like just setting targetFound = true instead of actually
calculating it).

MeleeWeapon
Checks for a health component in the hit collider's parent so it works for
trains.
Was worried this might make players not take damage because their parent
is the train, but so far, the player takes damage even when the train is
their parent.