NavMesh Creation and Triangulation
Added a number of classes to create and triangulate a navmesh

NavGraph
Main class that handles creating and triangulating the mesh
A number of points are given to determine the outer bounds of the mesh.
These points are assumed to be a concave polygon with each point given
being a point on the edge of the polygon.
The polygon is triangulated using an ear clipping algorithm.
The triangulated polygon is then passed through delaunay triangulation to
flip edges to make more uniform triangles if needed.

Helper classes
Edge
Helper class that defines what an edge is with some helper methods that
test is an edge intersects another edge

HalfEdge
A different way to store information about triangles. Used to perform
Delunay triangulation that aims to flip edges to get more uniform
triangles

Triangle
A class to store info about triangles (points, edges)
Has helper methods to determine if a point is inside of a triangle or if 2
triangles are intersecting.

Vertex
For use with triangles.
Stores its position and whether the point is reflex or convex and also if
it is an ear vertex (for ear clipping triangulation)

Problems
Points are in world coordinates still. They need to be converted to either
local coordinates or tied to transforms so the points move with the train.
Haven't made support to cut shapes out of the navmesh. Needed so agents
don't try to walk through solid objects