Convert Triangles to Nodes and Calculate Neighbours
NavGraph
CreateNodes() - a method to create nodes from triangles. The nodes are
used in A* pathfinding to hold the data the algorithm needs. It iterates
through the triangles and makes nodes coinciding with the transforms at
the vertices. Only makes nodes at unique vertices. As it creates the
nodes, it updates the neighbours of each node. This is the main reason for
this conversion procedure. A* needs to know about neighbours and the
trianlge implementation couldn't tell all of the neighbours.

FindPath() - the method that calculates paths. Uses A*. Uses a hashset and
heap structure as effeciency improvements. Not functional as it required
the neighbours of each node.
Still missing some implementation of the calculation of distance between
nodes.

ReconstructPath() - takes the path calculated by FindPath() and turns it
into a list of positions that an agent can follow.

Node
This is the container for the data A* needs.

Vertex
Now keeps track of the index of the transform that is associated with it

Heap
Generic heap structure to store the nodes for A* optimizations. The heap
stores the nodes in order as they are inserted so they don't need to be
sorted when they are accessed.