Path Calculating
NavGraph
CreateGraph()
I was trying to only add neighbours when a new node was created because I
thought the other neighbours would be duplicates. Turns out you can have
a triangle with all 3 nodes already evaluated, but have them not be neighbours with one
another. This is due to the order the ear-clipping algorithm works. In
certain shapes, it creates triangles on opposite sides of the shape. This
leaves the interior triangles as the last created. These triangles were
the ones that had the missing neighbours.

GetClosestNode()
For use with the agents finding paths. They give their current positions
and this returns the nearest node (the pathfinding wants nodes). It is not
necessarily the closest node. It is technically the closest node of the
triangle the given point is inside of. It is possible that there is a
closer point outside of that triangle. This is probably better 99% of the
time and even that 1% won't make too much of a difference.

FindPath()
Now finds existing nodes instead of creating new ones (was jsut a
placeholder, didn't actually function)
Exit condition was inside the wrong curly brace.

Added some testing for the paths and a gizmo to draw the path

AddNeighbours(n1, n2)
Add neighbours 2 at a time instead of individually. Figured this would be
easier because 2 neighbours are always evaluated at once. Adding
neighbours resizes the neighbours array so that now only needs to happen
once. Distances between neigbours is precalculated because I thought that
would help. Turns out knowing the distance between and node and the GOAL
node is more important (and the goal node is not usually a neighbour) so
this doesn't help.

GetDistBetween(n)
Used the afformentioned distance between neighbours because that's what I
thought was needed. Not used.
Not removing these because I'm not 100% sure that I don't need them; they
may be useful later.

CompareTo()
For sorting the heap. Sort by smaller fCost. Tie breakers by lower hCost

IsPointInTriangle()
Now the vertices are considered inside the triangle. Before, if the point
was exactly the same as a vertice, it didn't count as inside the triangle.
Now it does. (Actually, before, one of the points counted as inside, but the other 2 didn't)