Triangulation of Interior Points
Now the nav graph can have interior points that also get triangulated.
Before, all given points were assumed to be a part of the hull. This isn't
useful right now, but it will be useful later to be able to create holes
in the graph where there are unwalkable areas.