Constrained Delaunay
Added functionaity for constrained delaunay triangulation.
What it really does is allow you to specify a space inside the mesh that
is cut out.
You give it points which are interpreted as a series of edges.
Then it finds edges that intersect those constraint edges. It flips edges
so that there aren't any intersections.
Then it destroys the interior triangles.
There's a link to the article with pictures at the top of the file.

Fixed a bug in A*
Was calculating the cost between 2 points and then casting it to an int
(truncating). Changed to rounding to an int.
I'm not sure if rounding is good enough or if the calculations need to be
floats instead.
e.x. triangle with sides 15.4, 2.4, 17.6
15.4 + 2.4 = 17.8
17.6
Rounding
15+2 = 17
18
17.8 > 17.6
17 < 18
Rounding can give the wrong answer.

Updated the visual debugging of the path to include the path from the
current point to the first point. GetPath() omits the start node in the
path it returns (because that is where the agent currently is).