Expanded Mesh
Expanded the mesh to the third car.
It is all one mesh attached to the
second car at the moment. I'm not sure if one mesh is good or if each car
should have separate meshes with jumps between them.

Created a Hole class to store info about the holes to cut out of the mesh
(just the array of points that determines the hole). This allows me to
create any number of holes in the mesh from the inspector.
Reworked some of the mesh building code to work with multiple holes
instead of just the one array of inner points that was implied to be the
only hole.

Created the points to determine the mesh for the third car.
Cut a hole around the cargo container on that car too.
Renamed the mesh points so they are more descriptive.
Transforms of these extra points are children of the third train so they
follow with where that car is.

Updated debugging info
The mesh visuals now move to where the mesh currently is instead of just
showing where the mesh was first created.