Enemy Pathing
The enemy now pathfinds to its target when it boards the train.

When the enemy boards the train, it tries to find a path. If it finds a
path, it uses that path to move to its target. Otherwise, it just moves in
a straight line to its target.

Had to change some of the code that checks if a point is inside of a
triangle. It was checking every triangle in listOfTris. The problem was
that those triangles only had their original positions from when they were
first created (so they didn't update as the train and mesh moved). The
workaround is to create a new triangle and take the info from the triangle
to find where the corresponding tranform currently is, because vertecies
of the triangles have a reference to their associated transforms.

Also added a reference to the NavGraph to each train. So when an agent
boards the train, they check that train for the reference to the navMesh
to use.