Boarding Train State Logic
Enemy
Renamed target to moveTarget. It's where the enemy is trying to move
First pass at state transition. Most of the obvious ones. Helper methods
inline to keep them readable. This is a C#6 thing so might need to upgrade
the unity version for them to work.
Removed some old comments and the older version of the state based AI
Helper methods for the different states. Some don't do anything yet.
InterceptTrain() now keeps track of the boarding link it's trying to move
to. This is used by the BoardTrain state. Uses a tuple to make things
easy. This is a C#7 thing so it might not work.
BoardTrain() helper method. Now the AI needs to tell the enemy to board
the train instead of it just running into the trigger sphere and being
teleported aboard.

NavGraph
Noticed the pathfinding was looping through all triangles to see if the
start and end points were pathable, but then not keeping track of what
triangle they found the points in and then looping through all triangles
again. So I made them cache what triangle they are found in to be used
later. GetClosestNode(pos, tri) can be given a triangle instead of
searching all triangles.

AttackHostile
comments

BoardTrain
Changed variable names

MoveToRZPoint
There's no navmesh on the ground (yet?) so changed the implementation to
just move towards the target in a straight line.
When you leave the state, set a flag so you don't use an old position

SearchForTrainIntercept
enemy.InterceptTrain() already has some safety measures so I jsut let it
do the work. It automatically retries itself if the train isn't ready to
be boarded.
The only smart thing this state does is maybe do something if it takes too
long to intercept the train (probably because the train never moved at the
start of the level)

Train
GetBoardingLocationsInTime()
Now returns the array of boarding links in addtion to their positions in
time t.
Needed for Boarding state
Uses tuples which may or may not work in Unity depending on what version
of C# it can support.