AttackState, SearchOnTrain, and Unity Version Update
Updated to a new version of Unity. It uses a new version of C# so more
features.

Enemy
More helper fields and methods for use with the AI
Now tracks when its path ends
GetSightDistance() uses raycasts to tell if it can see something. If it
can, it returns how far away that thing is.

AttackHostile
Tries to attack its hostile target.
If it can see it and is inside attack range, attack it
If it can see, but outside the range, move closer
Has a min and max attack ranges. It tries to stay inside this range to be
more fluid. So that there's not just one range that might lead to the
enemy stutter stepping. One step takes it inside the range and it stops,
then next frame the target moves farther away so you take one more step.
Instead, you move until within the smaller range, then don't move until
you fall outside the outer range. Then move into the inner range again.

If you lose sight of target, keep moving to the last position.
Every frame, if you can see the target, you calculate a path to them. This
path stays if you lose sight of them so you will keep moving until you see
them again and a new path is generated.
If you reach the end of this path and still can't see the target, you swap
to a new state.

InvestigateHostile
Not sure if this is necessary or if the functionality from AttackHostile
is good enough.
Might still need this for more complex investigation than just going to
the last place they were spotted.

SearchForTargetsOnTrain
Get the set of all friendlies on the train
Iterate through them.
Find all the friendlies you can see. Sort by closest
The closest one you can see is your target.

Train
Store the set of all friendlies and hostiles on the train
I have some sets for each individual train car. Not sure if I need to know
the contents of each car or just the train as a whole.
Use BoardedTrain() to add yourself to the necessary sets.
Use LeaveTrain() when you jump off the train (onto the ground). Not when
swapping from one train car to another.

Nothing really uses these boarding/leaving methods yet. They currently just check if they
collide with an object with the "Train" tag and set that to be their
parent. Not really sure how best to make this work so that you can't
forget to add or remove yourself from these sets.