Train Manager and Intercept
TrainScene
Named the trains and their boarding links better names for debugging

Enemy
Hooked up the new state machine logic for testing
The diff doesn't show it nicely, but basically I redid the movement logic
in fixedUpdate. If you minimize the old versions, it's actually reasonbly
readable.
Missing the correct looking logic. Like looking where you're walking.
Intercept
Moved the intercept logic to Train Manager (new class).
It still tries to find a place and time where it can board. If it can't
find an intercept, it returns the closest it can. It sets some flags
saying it failed so the AI can check again when more trains enter the
tracks (this is only an issue in like the first 10 seconds of the level).

Enemy + Player
They tell the train they boarded it when they collide with it. It adds
them to sets of objects aboard the train.

Various states
Added debug info when entering states

BoardTrain
Fixed a bug. Only ever checked the target boarding link the agent had at
runtime. This target link can change, but wasn't being checked.

Train + TrainManager
Each train car is separate. They don't even know about one another. They
just all travel on the same path at the same speed so they look like
they're attached.
TrainManager handles the train as a whole. It knows the order of the
trains.

Train
CanBoard
The AI would ask the train if it could board and give a callback that was
called when it was able to board the train.
TrainManager used the CanBoard to see if trains were ready for its
intercept method. The problem was that both callbacks were called at the
same time and the AI methods would run first and use TrainManager methods
that also needed this info, causing problems. NeedEarly flag is for the
TrainManager to have a slight edge here so it has the right info when the
AI calls its methods.
This fix doesn't feel right.

TrainManager
Uses singleton pattern. Only 1 in the scene. Any class can find it in the
scene when it needs to know stuff about the train as a whole.
Moved the intercept logic here from Enemy
TrainManager knows about all of the trains and their order so the
intercept method can check multiple trains.