Attacking Movement
Enemy
wantToMove - you have to want to move to be able to move. This is to allow
the AI to stand still on the train.

GetSightDistance()
Fixed the raycasting. The layer mask didn't work how I thought/wanted it
to. I thought if I passed in the player's layer, it would only return if
it hit the player. Hitting a wall wouldn't return a wall. Instead it did
return a hit for whatever it hit first.
That isn't how layer masks work.
So now it just checks to see if it hit a player and only returns the dist
if it did

Removed some unused methods.

SetMoveTarget()
The pathing doesn't return a path if the start and end are in the same
triangle. Made some workaround for that.
TargetMarker was just a visual debugging tool to see where stuff was
headed. Now it works as the end of the path.
When finding a path, you put in vector3s, but because the train is moving,
the path gives you back transforms that move with the train. But the last
point isn't the position you wanted to go to, but the nearest one of these
transform nodes. TargetMarker is how I can make a node that I can move to
the exact destination position that will move with the train.

NavGraph
In an earlier commit, I tried to speed up checking what triangle the start
and end points were in. (It checked every trianlge already so I just
remembered which triangle they were in so it didn't have to check again).
However, the triangle I gave it only had positions, but not the indicies
which it also needs. So I tracked the correct indicies as well.

AttackHostile
Diff got really confused with this one.
Created a precondition for tick that checks if you're too close to your
target. If you are, don't move.

SearchForTargetsOnTrain
Stand still so you don't fall off the train right when you board.

Translucent Mat
used to see the targetMarker for debugging, but not see it too much.