Cleanup Enemy Class
Enemy
Removed a bunch of code that is no longer needed. Most of it hasn't been
used in a long while anyway.
This means I've moved completely over to the new state based AI.

Fixed some bugs
In AttackHostile, it was possilbe for the enemy to get stuck. If they were
in between their preferred range and outer range, they want to stand
still (expected behaviour). If they lose sight of the player, they are
supposed to move to where they last saw the player to investigate. So if
they were in this sweet spot when they lost sight, they wouldn't move.
Added a method to Enemy so they can try to start moving again without
having to try to make a new path (their old path is where they last saw
the player). .....Actually, now that I type it out, that isn't necessarily
true. Their last path is where the player was when they entered this sweet
spot. The player could enter their view from the right side, get in the
sweet spot, walk around them and leave their sight on the left and the AI
will search for them to the right.....

Somewhat related, if the agent tried to get a path where the start and end
of the path is in the same navmesh triangle, it returns an empty array.
The code that tells when the path ended runs when you reach the last point
in the path and that point is removed from the path. If the path is empty,
nothing gets removed.
Long story short, when you make a path, if the path length is 0, set
pathEnded to true.

Added a move visible material to the player's bullets.

Added a death state.
Doesn't do anything, but it keeps other states from
running when the enemy is dead.