Enemy Spawning
Made a script to handle spawning enemies as you play.
It spawns an enemy every x seconds off to one side of the train.
It uses the train's position along the path to spawn enemies. It spawns
them 15 steps ahead of where the train is and on the opposite side of the
origin from where the train is.