Stats Manager
Added a stats manager class.
It's a singelton that stores base values for common stats.
Use it to increase move speed of ALL entities instead of having to tweak
each of them individually
Currently controls player and enemy move speed and projectile speed

Added a kill box that follows behind the camera to clean up enemies that
lag behind.

Swapped the move speed of the player and enemies to look up base speeds
from the manager

Spawner now tracks how many enemies are alive and doesn't spawn more if
there are too many. Most of the time, the enemies lag behind and get
killed by the kill box so I don't hit the cap.