Plundering
Enemies can now board the train and path to a stockpile to steal loot.
They can also path to a borading link to exit the train.

Enemy:
Decreased how close you need to be to a node for it to count as visited
for the path. Agents were getting stuck right on top of the node they were
trying to get to.
Changed how the Enemy give Disembark info

NavGraph
Had a bug. Tried to do some debugging.
Turns out the bug was I moved some of the waypoints to a different y. That
completely breaks the navigation. And the errors I get don't even point
towards that being the issue.

DisembarkTrain
This is the first time actually testing it. What was there already was
just mostly pseudocode bc I couldn't test it.
It's not quite right, but it's alright for now.
It only checks for boarding links on the train the enemy originally tried
to board on, not all trains or even the train they're currently on.

PlunderTrain
First time hooking up and testing this too.
All of the < and > were backwards and it took longer than I'd care to
admit to fix it.