Global Stockpile Searching, Fixed intercept bug
Enemy
Fixed a bug: AI was using the wrong move speed to check if it could
intercept the train. Speed is now updated before the AI starts instead of
after
State change from moving to train back to search. This is if you didn't
find an intercept so you were just going close. Now when you get to the
'close enough' position, you state change to try to intercept again.
Removed some code that seemed either wrong or redundant or just boiler
plate code? Not sure why it was the way it was.

BoardTrain
Just added some notes

SearchOnTrain
Now check for stockpiles on all trains instead just one
Maybe you should check your train first, then other trains if you don't
find anything on yours? But there might be times when the stockpile on the
next train over is closer than one on your train so this is fine

SearchForIntercept
reset hasTrainRZPoint because this state is trying to find that point and
having that be true is the way out of this state. This makes it so I can
use this state multiple times

TrainManager
Method to get all stockpiles