Closest Intercept
Changed to calculate the closest intercept, not the first
Old structure prioritized the first train. This version prioritizes the
shortest time to intercept
New version is mostly better. Some edge cases have the old version be
better. Not really sure why. They chould check the same points, just in a
different order.

You now try to find a new intercept halfway to your destination if you
intercept was a failure instead of when you get really close.
This looks smoother.

There's a weird 'bug' with which trains you search for intercepts with.
You can set a certain number of trains to try to intercept instead of the
whole train.
But it doesn't work as intended if you fail to find an intercept the first
time, because even though you failed, you still find a near enough point.
This can change your target train. When you then try to find a new
intercept, you start from this possibly new target train instead of the
original which may be unintended behaviour.
I'm not even sure if being able to only search for intercepts on certain
trains is behaviour the game even needs. When would I not want to search
for intercepts for the whole train?